chore: avoid circular dependency Circular dependencies dist/utils.js -> dist/mutation-observers.js -> dist/aom.js -> dist/utils.js dist/utils.js -> dist/mutation-observers.js -> dist/utils.js
